+++ iOS Example/Sources/ObjCChartView.h
@@ -1,0 +1,18 @@
+//
+//  ObjCChartView.h
+//  iOS Example
+//
+//  Created by Andreas Neusüß on 12.06.20.
+//  Copyright © 2020 Anerma. All rights reserved.
+//
+
+#import <UIKit/UIKit.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+// Mock object for demonstrating Objective-C interface.
+@interface ObjCChartView : UIView
+
+@end
+
+NS_ASSUME_NONNULL_END